+++ pytest_ipynb/plugin.py
@@ -1,64 +1,37 @@
-import pytest
-import os, sys
-import warnings
-from pprint import pprint
-
+import pytest, os, sys, nbformat
+from queue import Empty
 from nbformat import NotebookNode
 import IPython
+from pprint import pprint
 
-try:
-    from exceptions import Exception, TypeError, ImportError
-except:
-    pass
-
+# XXX: runipy is no longer being maintained! https://github.com/paulgb/runipy
+# but it works perfectly fine.
+# In the future need to figure out how to move to new api
+# https://nbconvert.readthedocs.io/en/latest/execute_api.html
+# this implementation seems to do it in a modern way:
+# https://github.com/ldiary/pytest-testbook
 from runipy.notebook_runner import NotebookRunner
 
 wrapped_stdin = sys.stdin
 sys.stdin = sys.__stdin__
 sys.stdin = wrapped_stdin
-try:
-    from Queue import Empty
-except:
-    from queue import Empty
-
-# code copied from runipy main.py
-with warnings.catch_warnings():
-    try:
-        from IPython.utils.shimmodule import ShimWarning
-        warnings.filterwarnings('error', '', ShimWarning)
-    except ImportError:
-        class ShimWarning(Warning):
-            """Warning issued by iPython 4.x regarding deprecated API."""
-            pass
-
-    try:
-        # IPython 3
-        from IPython.nbformat import reads, NBFormatError
-    except ShimWarning:
-        # IPython 4
-        from nbformat import reads, NBFormatError
-    except ImportError:
-        # IPython 2
-        from IPython.nbformat.current import reads, NBFormatError
-    finally:
-        warnings.resetwarnings()
 
 class IPyNbException(Exception):
     """ custom exception for error reporting. """
 
 def pytest_collect_file(path, parent):
-    if path.fnmatch("test*.ipynb"):
-        return IPyNbFile(path, parent)
+    if path.fnmatch("test*.ipynb"): return IPyNbFile(path, parent)
 
 def get_cell_description(cell_input):
     """Gets cell description
 
-    Cell description is the first line of a cell,
-    in one of this formats:
+    Cell description is the first line of a cell, in one of this formats:
 
     * single line docstring
     * single line comment
     * function definition
+    * %magick
+
     """
     try:
         first_line = cell_input.split("\n")[0]
@@ -70,16 +43,15 @@ def get_cell_description(cell_input):
 
 class IPyNbFile(pytest.File):
     def collect(self):
-        with self.fspath.open() as f:
-            payload = f.read()
-        self.notebook_folder = self.fspath.dirname
-        try:
-            # Ipython 3
-            self.nb = reads(payload, 3)
-        except (TypeError, NBFormatError):
-            # Ipython 2
-            self.nb = reads(payload, 'json')
+        with self.fspath.open() as f: payload = f.read()
+        self.nb = nbformat.reads(payload, 3)
+
+        # kernel needs to start from the same dir the ipynb is in
+        notebook_dir = self.fspath.dirname
+        cwd = os.getcwd()
+        if cwd != notebook_dir: os.chdir(notebook_dir)
         self.runner = NotebookRunner(self.nb)
+        os.chdir(cwd)
 
         cell_num = 1
 
@@ -110,7 +82,6 @@ class IPyNbCell(pytest.Item):
         cell_description = get_cell_description(cell.input)
         nodeid = parent.nodeid + "::" + f"cell {cell_num:2d}"
         if cell_description: nodeid += " " + cell_description[0:40]
-
         super(IPyNbCell, self).__init__(name, parent, nodeid=nodeid)
 
         self.cell_num = cell_num
@@ -121,7 +92,7 @@ class IPyNbCell(pytest.Item):
     # use run_cell from runipy/notebook_runner.py as a base for a better runtest
     # implementation than the original pytest-ipynb, to include reporting output streams
     def runtest(self):
-        self.kc = self.parent.runner.kc
+        kc = self.parent.runner.kc
         cell = self.cell
 
         if ("SKIPCI" in self.cell_description) and ("CI" in os.environ):
@@ -133,20 +104,26 @@ class IPyNbCell(pytest.Item):
         if self.cell_description.lower().startswith("fixture") or self.cell_description.lower().startswith("setup"):
             self.parent.fixture_cell = self.cell
 
-        self.kc.execute(cell.input, allow_stdin=False)
-        reply = self.kc.get_shell_msg()
+        kc.execute(cell.input, allow_stdin=False)
+        # XXX: the way it's currently implemented there doesn't seem to be a
+        # point in handling a timeout situation here, since raising an exception
+        # on timeout breaks the rest of the tests. The correct way to do it is to
+        # send interrupt to the kernel, and then report timeout, so that the
+        # rest of the tests can continue.
+        reply = kc.get_shell_msg()
+
         status = reply['content']['status']
         traceback_text = ''
         if status == 'error':
             traceback_text = 'Cell raised uncaught exception: \n' + \
                 '\n'.join(reply['content']['traceback'])
 
+        # extract various outputs and streams
         outs = list()
-        # XXX: we really only care for pyout and stdout streams below, so the following can be pruned a lot
         timeout = 20
         while True:
             try:
-                msg = self.kc.get_iopub_msg(timeout=timeout)
+                msg = kc.get_iopub_msg(timeout=timeout)
                 if msg['msg_type'] == 'status':
                     if msg['content']['execution_state'] == 'idle':
                         break
@@ -159,129 +136,72 @@ class IPyNbCell(pytest.Item):
             content = msg['content']
             msg_type = msg['msg_type']
 
-            # IPython 3.0.0-dev writes pyerr/pyout in the notebook format
-            # but uses error/execute_result in the message spec. This does the
-            # translation needed for tests to pass with IPython 3.0.0-dev
-            notebook3_format_conversions = {
-                'error': 'pyerr',
-                'execute_result': 'pyout'
-            }
-            msg_type = notebook3_format_conversions.get(msg_type, msg_type)
-
             out = NotebookNode(output_type=msg_type)
 
             if 'execution_count' in content:
-                cell['prompt_number'] = content['execution_count']
                 out.prompt_number = content['execution_count']
 
             if msg_type in ('status', 'pyin', 'execute_input'):
                 continue
             elif msg_type == 'stream':
                 out.stream = content['name']
-                # in msgspec 5, this is name, text
-                # in msgspec 4, this is name, data
-                if 'text' in content:
-                    out.text = content['text']
-                else:
-                    out.text = content['data']
-            elif msg_type in ('display_data', 'pyout'):
+                if 'text' in content: out.text = content['text']
+            # execute_result == Out[] (_)
+            elif msg_type in ('display_data', 'execute_result'):
                 for mime, data in content['data'].items():
                     try:
                         attr = self.MIME_MAP[mime]
                     except KeyError:
-                        raise NotImplementedError(
-                            'unhandled mime type: %s' % mime
-                        )
+                        raise NotImplementedError(f'unhandled mime type: {mime}')
 
-                    # In notebook version <= 3 JSON data is stored as a string
-                    # Evaluation of IPython2's JSON gives strings directly
-                    # Therefore do not encode for IPython versions prior to 3
-                    json_encode = (
-                            IPython.version_info[0] >= 3 and
-                            mime == "application/json")
-
+                    json_encode = (mime == "application/json")
                     data_out = data if not json_encode else json.dumps(data)
                     setattr(out, "data_type", attr)
                     setattr(out, "data", data_out)
-            elif msg_type == 'pyerr':
+            elif msg_type == 'error':
                 out.ename = content['ename']
                 out.evalue = content['evalue']
                 out.traceback = content['traceback']
-            elif msg_type == 'clear_output':
-                outs = list()
-                continue
+            elif msg_type == 'clear_output': pass
+                # ignore
+                #outs = list()
+                #continue
             else:
-                raise NotImplementedError(
-                    'unhandled iopub message: %s' % msg_type
-                )
+                raise NotImplementedError(f'unhandled iopub message: {msg_type}')
             outs.append(out)
-        #cell['outputs'] = outs
+        #pprint(outs)
 
         if status == 'error':
-            # Extract all output streams, so that we can display them in the exception
+            # Get all output streams, so that we can display them in the exception
             pyout = []
             stdout_data = ''
+            stderr_data = ''
             for out in outs:
-                if out.output_type == 'pyout':
+                if out.output_type == 'execute_result':
                     # text, html, json (others are binary)
                     if out.data_type in ["text", "html", "json"]:
                         pyout.append(out.data)
                     else:
                         pyout.append(f"[{out.data_type} object]")
-                elif 'stream' in out and out.stream == 'stdout':
-                    stdout_data = out.text
-            #pprint(outs)
-            #pprint(pyout)
+                elif 'stream' in out:
+                    if out.stream == 'stdout':
+                        stdout_data = out.text
+                    elif out.stream == 'stderr':
+                        stderr_data = out.text
 
-            raise IPyNbException(self.cell_num, self.cell.input, "\n".join(pyout), stdout_data, traceback_text)
-
-
-    # this is the old implementation, which needs to be merged with the new one
-    # if we were to support all the previous features.
-    def runtest2(self):
-        kc = self.parent.runner.kc
-        self.kc = kc
-        cell = self.cell
-
-        # must not restart kernel for each cell! (XXX: needs to be made configurable for those who want it to be restarted).
-        #self.parent.runner.km.restart_kernel()
-
-        if self.parent.notebook_folder:
-            kc.execute(f"import os; os.chdir('{self.parent.notebook_folder}')")
-
-        if ("SKIPCI" in self.cell_description) and ("CI" in os.environ):
-            pass
-        else:
-            if self.parent.fixture_cell:
-                kc.execute(self.parent.fixture_cell.input, allow_stdin=False)
-            msg_id = kc.execute(self.cell.input, allow_stdin=False)
-            if self.cell_description.lower().startswith("fixture") or self.cell_description.lower().startswith("setup"):
-                self.parent.fixture_cell = self.cell
-
-            timeout = 20
-            while True:
-                try:
-                    msg = kc.get_shell_msg(block=True, timeout=timeout)
-                    if msg.get("parent_header", None) and msg["parent_header"].get("msg_id", None) == msg_id:
-                        break
-                except Empty:
-                    raise IPyNbException(self.cell_num, self.cell.input, "no output?",
-                                         "no stdout",
-                                         "Timeout of %d seconds exceeded executing cell: %s" % (timeout, self.cell.input))
-
-            reply = msg['content']
-            if reply['status'] == 'error':
-                raise IPyNbException(self.cell_num, self.cell.input, "nooutput?", "nostdout?", '\n'.join(reply['traceback']))
+            raise IPyNbException(self.cell_num, self.cell.input, "\n".join(pyout),
+                                 stdout_data, stderr_data, traceback_text)
 
     def repr_failure(self, excinfo):
         """ called when self.runtest() raises an exception. """
         if isinstance(excinfo.value, IPyNbException):
             return "\n".join([
-                "*** Failed cell %d: ***\n\n"
-                "*** Input ***\n%s\n\n"
-                "*** Output ***\n%s\n\n"
-                "*** STDOUT ***\n%s\n\n"
-                "*** Traceback ***\n%s\n\n\n\n" % excinfo.value.args,
+                "====== Failed cell %d ======\n\n"
+                "====== input ======\n%s\n\n"
+                "====== output ======\n%s\n\n"
+                "====== stdout ======\n%s\n\n"
+                "====== stderr ======\n%s\n\n"
+                "====== traceback ======\n%s\n\n\n\n" % excinfo.value.args,
             ])
         else:
             return "pytest plugin exception: %s" % str(excinfo.value)